refactor(macos): split MainView into focused subcomponents
Break the 1125-LOC MainView.swift into a thin composition root plus
seven focused extension files under apps/macos/cubelite/cubelite/Views/:

- MainView.swift (201 LOC)               — state + body wiring + previews
- MainView+Toolbar.swift (68 LOC)        — toolbar + logs button
- MainView+Sidebar.swift (298 LOC)       — left pane (clusters / namespaces)
- MainView+ContentColumn.swift (36 LOC)  — middle column resource type list
- MainView+DetailArea.swift (156 LOC)    — right pane router + browser
- MainView+ConfigLoader.swift (142 LOC)  — kubeconfig + namespace loaders
- MainView+ResourceLoader.swift (185 LOC) — per-namespace resource fetch
- MainView+CrossClusterLoader.swift (103 LOC) — All Clusters aggregation

Pure structural split — no behavior changes, no API changes, no logic
tweaks. State property visibility relaxed from 'private' to internal so
cross-file extensions can observe and mutate it; semantics are identical.
Existing TODO markers reference issues #104 and #106 for pre-existing
issues uncovered (force-cast and throwaway AppSettings()) — not fixed
here per the Bug Discovery Workflow.

Closes #125
